+++ src/Lib/Highlight.ts
@@ -1,8 +1,6 @@
 import { GeometryType } from '../Type/Geometry.type';
 
 export class Highlight {
-  static instance: Highlight;
-
   public data: Map<GeometryType, Map<number, { [key: string]: any }>> = new Map([
     [GeometryType.rect, new Map()],
     [GeometryType.text, new Map()],
@@ -16,10 +14,12 @@ export class Highlight {
     return this.data.get(type);
   }
 
-  static from(): Highlight {
-    if (!this.instance) {
-      this.instance = new Highlight();
-    }
-    return this.instance;
+  public reset() {
+    this.data = new Map([
+      [GeometryType.rect, new Map()],
+      [GeometryType.text, new Map()],
+      [GeometryType.image, new Map()],
+      [GeometryType.path, new Map()],
+    ]);
   }
 }